+++ bir/Cargo.toml
@@ -4,11 +4,16 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bincode = { version = "1.3" }
 http = { version = "1.1" }
 http-body-util = { version = "0.1" }
 hyper-util = { version = "0.1" }
 hyper = { version = "1.1", features = ["client", "http1", "server"] }
-tokio = { version = "1", features = ["full"] }
+tokio = { version = "1.38", features = ["full"] }
+tokio-rustls = { version = "0.24" }
+tokio-util = { version = "0.7" }
+tracing = { version = "0.1" }
+tracing-subscriber = { version = "0.3" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 reqwest = { version = "0.12.9", features = ["json"]}
